I think I fixed the first error from Clingers report.
There wasnt an if statement covering the case when an Enemy does
a move in the updateBoard funtion for some reason. I added it but
am having trouble testing it.

Currently when we go to make a move we are just looking for the first
open move. This leads us to enter a loop where we are just moveing
back and forth between two open positions. And since we are playing
ourself when testing the two sides never meet, so i can't replicate the
case where Clinger's error occurred. If we alter the way we make moves to
favore moving toward the enemy, this might work, but i think this is what Dan
was working towards so I'm committing the current fix now.